Add new generation features and improve maze renderer
The maze now contains randomly placed patches of grass that are placed
using a simplex noise map.
The MazeGenerator now has fail-safes in order to ensure that a map can
always be generated from any given seed. If a certain seed is unable to
generate any proper map due to to patches of grass making it impossible
to generate a functioning map then it will modify the seed for the noise
map to create a different result.
Furthermore the generator now ensure that an entrance and exit was placed
properly.

The MazeDrawer now caches tiles in order to draw mazes faster. This
becomes especially noticeable when drawing lots of debug tiles.

The Interface has a few more buttons added in order to be able to set a
seed, change the noise and switch to a view of the map with debug tiles.